+++ scripts/ci/wasm-report-common.sh
@@ -5,17 +5,19 @@ wasm_report_default_canisters() {
         default_empty \
         default_empty_metrics \
         one_entity_dynamic_query \
+        one_entity_reachable_operations \
         one_entity_typed_query \
         one_entity_sql_query \
         request_future_scale \
         ten_entity_typed_query \
+        ten_entity_reachable_operations \
         sql_perf \
         sql
 }
 
 wasm_report_canister_is_maintained_subject() {
     case "$1" in
-        default_empty|default_empty_metrics|one_entity_dynamic_query|one_entity_typed_query|one_entity_sql_query|request_future_scale|ten_entity_typed_query|sql_perf|sql)
+        default_empty|default_empty_metrics|one_entity_dynamic_query|one_entity_reachable_operations|one_entity_typed_query|one_entity_sql_query|request_future_scale|ten_entity_typed_query|ten_entity_reachable_operations|sql_perf|sql)
             return 0
             ;;
         *)